+++ CHANGELOG.md
@@ -25,7 +25,7 @@
 
 ### Updated
 
-- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.5.0-alpha.10`
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.5.0-alpha.12`
 - `gradle -> 8.1`
 
 ## [1.0.0-alpha.30] - 2023-03-05