+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### Updated
 
+- `gradle -> 8.1.1`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.5.0-alpha.16`
 
 ## [1.0.0-alpha.31] - 2023-04-28