+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Updated
 
+- `gradle -> 8.2`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.5.0-alpha.39`
 
 ## [1.0.0-alpha.32] - 2023-05-18