+++ CHANGELOG.md
@@ -14,8 +14,8 @@
 
 ### Updated
 
+- `gradle -> 8.3`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.5.0-beta.10`
-- `gradle -> 8.2.1`
 
 ## [1.0.0-alpha.32] - 2023-05-18